# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ContaNova ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: Listar clientes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: activo
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Crear cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteInput"
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Obtener cliente por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "404":
          description: No encontrado
    patch:
      operationId: updateCliente
      tags: [clientes]
      summary: Actualizar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteUpdate"
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Eliminar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado
  /clientes/{id}/resumen:
    get:
      operationId: getClienteResumen
      tags: [clientes]
      summary: Resumen de cuenta del cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resumen del cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteResumen"

  # ─── PROVEEDORES ─────────────────────────────────────────────────────────────
...
  /compras:
    get:
      operationId: listCompras
      tags: [compras]
      summary: Listar órdenes y facturas de compra
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, recibida, pagada, anulada]
        - name: proveedorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de compras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Compra"
    post:
      operationId: createCompra
      tags: [compras]
      summary: Crear orden/factura de compra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompraInput"
      responses:
        "201":
          description: Compra creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
  /compras/{id}:
    get:
      operationId: getCompra
      tags: [compras]
      summary: Obtener compra
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Compra con detalle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompraDetalle"
    patch:
      operationId: updateCompra
      tags: [compras]
      summary: Actualizar compra
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompraUpdate"
      responses:
        "200":
          description: Compra actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
    delete:
      operationId: deleteCompra
      tags: [compras]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # ─── CARTERA ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Cliente:
      type: object
      required:
        - id
        - nombre
        - tipoDocumento
        - numeroDocumento
        - activo
        - createdAt
      properties:
        id:
          type: integer
        nombre:
          type: string
        tipoDocumento:
          type: string
        numeroDocumento:
          type: string
        email:
          type: ["string", "null"]
        telefono:
          type: ["string", "null"]
        direccion:
          type: ["string", "null"]
        ciudad:
          type: ["string", "null"]
        activo:
          type: boolean
        saldoPendiente:
          type: ["number", "null"]
        createdAt:
          type: string
    ClienteInput:
      type: object
      required: [nombre, tipoDocumento, numeroDocumento]
      properties:
        nombre:
          type: string
          minLength: 1
        tipoDocumento:
          type: string
        numeroDocumento:
          type: string
        email:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        ciudad:
          type: string
    ClienteUpdate:
      type: object
      properties:
        nombre:
          type: string
        tipoDocumento:
          type: string
        numeroDocumento:
          type: string
        email:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        ciudad:
          type: string
        activo:
          type: boolean
    ClienteResumen:
      type: object
      required:
        - clienteId
        - totalFacturado
        - totalPagado
        - saldoPendiente
        - cantidadFacturas
      properties:
        clienteId:
          type: integer
        totalFacturado:
          type: number
        totalPagado:
          type: number
        saldoPendiente:
          type: number
        cantidadFacturas:
          type: integer
        ultimaCompra:
          type: ["string", "null"]

    # ─── PROVEEDORES ──────────────────────────────────────────────────────────
...
    CompraItem:
      type: object
      required: [id, productoId, descripcion, cantidad, precioUnitario, subtotal]
      properties:
        id:
          type: integer
        productoId:
          type: ["integer", "null"]
        descripcion:
          type: string
        cantidad:
          type: number
        precioUnitario:
          type: number
        subtotal:
          type: number
    CompraItemInput:
      type: object
      required: [descripcion, cantidad, precioUnitario]
      properties:
        productoId:
          type: integer
        descripcion:
          type: string
        cantidad:
          type: number
        precioUnitario:
          type: number
    Compra:
      type: object
      required:
        - id
        - numero
        - proveedorId
        - proveedorNombre
        - fecha
        - total
        - estado
        - createdAt
      properties:
        id:
          type: integer
        numero:
          type: string
        proveedorId:
          type: integer
        proveedorNombre:
          type: string
        fecha:
          type: string
        fechaVencimiento:
          type: ["string", "null"]
        subtotal:
          type: number
        impuesto:
          type: number
        total:
          type: number
        estado:
          type: string
          enum: [pendiente, recibida, pagada, anulada]
        notas:
          type: ["string", "null"]
        createdAt:
          type: string
    CompraDetalle:
      type: object
      required:
        - id
        - numero
        - proveedorId
        - proveedorNombre
        - fecha
        - total
        - estado
        - items
        - createdAt
      properties:
        id:
          type: integer
        numero:
          type: string
        proveedorId:
          type: integer
        proveedorNombre:
          type: string
        fecha:
          type: string
        fechaVencimiento:
          type: ["string", "null"]
        subtotal:
          type: number
        impuesto:
          type: number
        total:
          type: number
        estado:
          type: string
          enum: [pendiente, recibida, pagada, anulada]
        notas:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CompraItem"
        createdAt:
          type: string
    CompraInput:
      type: object
      required: [proveedorId, fecha, items]
      properties:
        proveedorId:
          type: integer
        fecha:
          type: string
          format: date
        fechaVencimiento:
          type: string
          format: date
        notas:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CompraItemInput"
    CompraUpdate:
      type: object
      properties:
        fecha:
          type: string
        fechaVencimiento:
          type: string
        estado:
          type: string
          enum: [pendiente, recibida, pagada, anulada]
        notas:
          type: string

    # ─── CARTERA ──────────────────────────────────────────────────────────────
...